refactor: split JS into modules
