+++ announcement.html
@@ -69,7 +69,7 @@
         <!-- Footer content will be loaded here by JavaScript -->
     </footer>
 
-    <script src="script.js"></script>
+    <script type="module" src="js/main.js"></script>
 
 </body>
 </html>
