+++ index.html
@@ -187,7 +187,7 @@
         <!-- Footer content will be loaded here by JavaScript -->
     </footer>
 
-    <script src="script.js"></script>
+    <script type="module" src="js/main.js"></script>
 
 </body>
 </html>
