+++ src/pages/blog/index.html
@@ -116,7 +116,7 @@
         <!-- Footer content will be loaded here by JavaScript -->
     </footer>
 
-    <script src="script.js"></script>
+    <script type="module" src="js/main.js"></script>
     <script>
         // Small script to show the update banner based on the main page's logic, if needed
         // Or just remove/comment out the banner if it shouldn't show here.
