+++ src/pages/blog/integration-flashcards/index.html
@@ -403,7 +403,7 @@
         }); // End DOMContentLoaded
     </script>
     <!-- Link to global script.txt if you have one for other functions -->
-    <!-- <script src="script.js"></script> --> <!-- Assuming you rename script.txt -->
+    <!-- <script type="module" src="js/main.js"></script> --> <!-- Assuming you rename script.txt -->
 
 </body>
 </html>
