+++ src/pages/blog/tck-guide-netherlands/index.html
@@ -155,7 +155,7 @@
         <!-- Footer content will be loaded here by JavaScript -->
     </footer>
 
-    <script src="script.js"></script>
+    <script type="module" src="js/main.js"></script>
 
 </body>
 </html>
