+++ tos.html
@@ -234,6 +234,6 @@
         <!-- Footer content will be loaded here by JavaScript -->
     </footer>
 
-    <script src="script.js"></script>
+    <script type="module" src="js/main.js"></script>
 </body>
 </html>
